Use a context for feature visibility and rename off Store
Replace the Zustand store with a context (synchronous, no post-mount lag),
caching the value by a ref so consumer identity stays stable without
unpacking each flag. Rename the provider/module from
*FeatureVisibilityStore* to *FeatureVisibility*.
